refactor: recorded array methods in a separate file
